tweak css and add vertices

+++ hello-world/shaders/fragment.glsl
@@ -1,4 +1,4 @@
-varying lowp vec4 color;
+varying highp vec4 color;
 
 void main() {
     gl_FragColor = color;
